// packages/payload/src/database/types.ts
export type FindOneArgs = {
  collection: string
  locale?: string
  req: PayloadRequest
  where?: Where
}